requirements.txt
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-storage